requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiohttp-cors==0.7.0
aiosignal==1.3.1
alembic==1.13.3
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
async-timeout==4.0.3
attrs==24.2.0
Brotli==1.1.0
cached-property==1.5.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
colorful==0.5.6
colorlog==6.8.2
contourpy==1.3.0
cryptography==43.0.1
cycler==0.12.1
distlib==0.3.8
filelock==3.16.1
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.9.0
gmpy2==2.1.5
google-api-core==2.21.0
google-auth==2.35.0
googleapis-common-protos==1.65.0
GPUtil==1.4.0
greenlet==3.1.1
grpcio==1.62.2
h2==4.1.0
h5py==3.11.0
hpack==4.0.0
huggingface-hub==0.25.2
hydra-core==1.3.2
hyperframe==6.0.1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
lightning==2.4.0
lightning-utilities==0.11.7
linkify-it-py==2.0.3
Mako==1.3.5
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.1
matplotlib==3.9.2
mdit-py-plugins==0.4.2
mdurl==0.1.2
memray==1.12.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
munkres==1.1.4
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
opencensus==0.11.3
opencensus-context==0.1.3
optuna==4.0.0
packaging==24.1
pandas==2.2.3
pillow==11.0.0
pkgutil_resolve_name==1.3.10
platformdirs==3.11.0
prometheus_client==0.21.0
proto-plus==1.23.0
protobuf==4.25.3
psutil==6.0.0
pyarrow==16.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyOpenSSL==24.2.1
pyparsing==3.1.4
PySide6==6.7.2
PySide6_Addons==6.7.2
PySide6_Essentials==6.7.2
PySocks==1.7.1
python-dateutil==2.9.0
pytorch-lightning==2.4.0
pytz==2024.1
pyu2f==0.1.5
PyYAML==6.0.2
ray==2.36.1
referencing==0.35.1
requests==2.32.3
rich==13.9.2
rpds-py==0.20.0
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.2
scikit-plot==0.3.7
scipy==1.11.4
setproctitle==1.3.3
shiboken6==6.7.2
six==1.16.0
smart-open==7.0.5
SQLAlchemy==2.0.35
sympy==1.13.3
tensorboard==2.18.0
tensorboard-data-server==0.7.0
tensorboardX==2.6.2.2
textual==0.82.0
threadpoolctl==3.5.0
timm==1.0.9
torch==2.4.1
torch-geometric==2.6.1
torchaudio==2.4.1
torchmetrics==1.4.2
torchvision==0.19.1
tornado==6.4.1
tqdm==4.66.5
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.2
uc-micro-py==1.0.3
unicodedata2==15.1.0
urllib3==2.2.3
virtualenv==20.21.0
Werkzeug==3.0.4
wrapt==1.16.0
yarl==1.13.1
zipp==3.20.2
zstandard==0.23.0
